+++ api/tests/certification/results/unit/application/certification-reports-controller_test.js
@@ -1,5 +1,5 @@
-import { certificationReportsController } from '../../../../../src/certification/course/application/certification-reports-controller.js';
-import { usecases } from '../../../../../src/certification/course/domain/usecases/index.js';
+import { certificationReportsController } from '../../../../../src/certification/results/application/certification-reports-controller.js';
+import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | certification-reports-controller', function () {
